+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3
+ - Intelligently map logstash events to the Timber.io log event JSON schema
+   (https://github.com/timberio/log-event-json-schema).
+
+## 1.0.2
+ - Add hostname to the payload being sent to Timber.
+
 ## 1.0.1
  - Drop the @version field since it is not useful to external systems